Moved http stuff to http.ccp/hpp
